+++ tests/ui/console_ui_test.js
@@ -6,17 +6,18 @@ import { ConsoleUI } from '../../lib/ui/console_ui.js';
 import { FakeConsole } from './fake_console.js';
 import { FakeProcess } from './fake_process.js';
 import { Configuration } from '../../lib/config/configuration.js';
+import { emptyRunnerCallbacks, withRunnerAndCallbacks } from '../support/runner_helpers.js';
 
 suite('Console UI', () => {
-  let fakeProcess, fakeConsole;
+  let fakeProcess, fakeConsole, console;
 
   before(() => {
     fakeProcess = new FakeProcess();
     fakeConsole = new FakeConsole();
+    console = new ConsoleUI(fakeProcess, fakeConsole);
   });
 
   test('can exit with error displaying the given message', () => {
-    const console = new ConsoleUI(fakeProcess, fakeConsole);
     console.exitWithError('a message');
     assert.that(fakeProcess.lastExitCode()).isEqualTo(ConsoleUI.failedExitCode());
     assert.that(fakeConsole.messages().length).isEqualTo(1);
@@ -31,16 +32,26 @@ suite('Console UI', () => {
     assert.areEqual(ConsoleUI.failedExitCode(), 1);
   });
 
-  test('prints start and end messages', async() => {
-    const console = new ConsoleUI(fakeProcess, fakeConsole);
+  test('prints initial configuration on start', async() => {
     const config = Configuration.current();
     await console.start(config, [], () => {});
-    assert.that(fakeConsole.messages().length).isEqualTo(8);
-    assert.that(fakeConsole.messages().at(0)).matches(/timer started/);
+    assert.that(fakeConsole.messages().length).isEqualTo(7);
     assert.that(fakeConsole.messages().at(2)).matches(/Starting Testy!/);
     assert.that(fakeConsole.messages().at(3)).matches(/Running tests in/);
     assert.that(fakeConsole.messages().at(4)).matches(/Fail fast/);
     assert.that(fakeConsole.messages().at(5)).matches(/Random order/);
-    assert.that(fakeConsole.messages().at(7)).matches(/timer ended/);
+  });
+
+  test('has a timer that ends at the end of the runner', async() => {
+    const callbacks = {
+      ...emptyRunnerCallbacks,
+      onFinish: runner => {
+        console.onRunnerFinish(runner);
+      },
+    };
+    await withRunnerAndCallbacks(callbacks, async runner => {
+      runner.run();
+    });
+    assert.that(fakeConsole.messages().at(-1)).isEqualTo('timer "Total time" ended');
   });
 });